% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_cvhierr.R
\name{coef.cv_hierr}
\alias{coef.cv_hierr}
\title{Get coefficient estimates from "cv_hierr" model object}
\usage{
\method{coef}{cv_hierr}(object, p = "opt", pext = "opt",
  penalty = NULL, ...)
}
\arguments{
\item{object}{A \code{\link{cv_hierr}} object}

\item{p}{vector of penalty values to apply to predictor variables.
Default is optimal value in cv_hierr object.}

\item{pext}{vector of penalty values to apply to external data variables.
Default is optimal value in cv_hierr object.}

\item{penalty}{(optional) regularization object applied to original model object, only needed
if p or pext are not in the original path(s) computed. See \code{\link{define_penalty}} for
more information on regularization object.}

\item{...}{pass other arguments to hierr function (if needed)}
}
\description{
Returns coefficents from 'hierr' model or refits model to estimate
coefficients not computed in path(s).
}
