% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_tune_xrnet.R
\name{coef.tune_xrnet}
\alias{coef.tune_xrnet}
\title{Get coefficient estimates from "tune_xrnet" model object}
\usage{
\method{coef}{tune_xrnet}(object, p = "opt", pext = "opt",
  penalty = NULL, ...)
}
\arguments{
\item{object}{A \code{\link{tune_xrnet}} object}

\item{p}{vector of penalty values to apply to predictor variables.
Default is optimal value in tune_xrnet object.}

\item{pext}{vector of penalty values to apply to external data variables.
Default is optimal value in tune_xrnet object.}

\item{penalty}{(optional) regularization object applied to original model object, only needed
if p or pext are not in the original path(s) computed. See \code{\link{define_penalty}} for
more information on regularization object.}

\item{...}{pass other arguments to xrnet function (if needed)}
}
\description{
Returns coefficents from 'xrnet' model or refits model to estimate
coefficients not computed in path(s).
}
