% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_xrnet.R
\name{coef.xrnet}
\alias{coef.xrnet}
\title{Get coefficient estimates from "xrnet" model object}
\usage{
\method{coef}{xrnet}(object, p = NULL, pext = NULL, penalty = NULL,
  ...)
}
\arguments{
\item{object}{A \code{\link{xrnet}} object}

\item{p}{vector of penalty values to apply to predictor variables.}

\item{pext}{vector of penalty values to apply to external data variables.}

\item{penalty}{(optional) regularization object applied to original model object, only needed
if p or pext are not in the original path(s) computed. See \code{\link{define_penalty}} for
more information on regularization object.}

\item{...}{pass other arguments to xrnet function (if needed)}
}
\description{
Returns coefficents from 'xrnet' model or refits model to estimate
coefficients not computed in path(s).
}
