% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvhierr.R
\name{cv_hierr}
\alias{cv_hierr}
\title{k-fold cross-validation for hierarchical regularized regression}
\usage{
cv_hierr(x, y, external = NULL, unpen = NULL, family = c("gaussian",
  "binomial"), penalty = definePenalty(), weights = NULL,
  standardize = c(TRUE, TRUE), intercept = c(TRUE, FALSE),
  loss = c("mse", "mae", "auc"), nfolds = 5, foldid = NULL,
  parallel = FALSE, control = list())
}
\arguments{
\item{x}{predictor design matrix of dimension \eqn{n x p}}

\item{y}{outcome vector of length \eqn{n}}

\item{external}{(optional) external data design matrix of dimension \eqn{p x q}}

\item{unpen}{(optional) unpenalized predictor design matrix}

\item{family}{error distribution for outcome variable}

\item{penalty}{specifies regularization object for x and external. See \code{\link{definePenalty}} for more details.}

\item{weights}{optional vector of observation-specific weights. Default is 1 for all observations.}

\item{standardize}{indicates whether x and/or external should be standardized. Default is c(TRUE, TRUE).}

\item{intercept}{indicates whether an intercept term is included for x and/or external. Default is c(TRUE, FALSE).}

\item{loss}{loss function for cross-validation. Options include:
\itemize{
   \item mse (Mean Squared Error)
   \item mae (Mean Absolute Error)
   \item auc (Area under the curve)
}}

\item{nfolds}{number of folds for cross-validation. Default is 5.}

\item{foldid}{(optional) vector that identifies user-specified fold for each observation. If NULL, folds are automatically generated.}

\item{parallel}{use \code{foreach} function to fit folds in parallel if TRUE, must register cluster (\code{doParallel}) before using.}

\item{control}{specifies hierr control object. See \code{\link{hierr.control}} for more details.}

\item{...}{list of additional arguments to pass to function \code{\link{hierr}}.}
}
\description{
k-fold cross-validation for hierarchical regularized regression \code{\link{hierr}}
}
