% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvhierr.R
\name{cvhierr}
\alias{cvhierr}
\title{k-fold cross-validation for hierarchical regularized regression}
\usage{
cvhierr(x, y, external, family = c("gaussian", "binomial"),
  penalty = definePenalty(0, 1), weights = NULL, type.measure = c("mse",
  "mae", "deviance"), nfolds = 5, parallel = FALSE, ...)
}
\arguments{
\item{x}{predictor design matrix of dimension \eqn{n x p}}

\item{y}{outcome vector of length \eqn{n}}

\item{external}{external data design matrix of dimension \eqn{p x q}}

\item{family}{error distribution for outcome variable}

\item{penalty}{specifies regularization object for x and external. See \code{\link{definePenalty}} for more details.}

\item{weights}{optional vector of observation-specific weights. Default is 1 for all observations.}

\item{type.measure}{loss function for cross-validation. Options include:
\itemize{
   \item mse (Mean Squared Error)
   \item deviance
   \item mae (Mean Absolute Error)
}}

\item{nfolds}{number of folds for cross-validation. Default is 5.}

\item{parallel}{Use \code{foreach} function to fit folds in parallel if TRUE, must register cluster (\code{doParallel}) before using.}

\item{...}{list of additional arguments to pass to function \code{\link{hierr}}.}
}
\description{
k-fold cross-validation for hierarchical regularized regression \code{\link{hierr}}
}
