% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definePenalty.R
\name{definePenalty}
\alias{definePenalty}
\title{Define regularization object for predictor and external data}
\usage{
definePenalty(penalty_type = 0, quantile = 0.5, penalty_type_ext = 1,
  quantile_ext = 0.5, num_penalty = 20, num_penalty_ext = 20,
  penalty_ratio = NULL, penalty_ratio_ext = NULL,
  user_penalty = NULL, user_penalty_ext = NULL,
  custom_multiplier = NULL, custom_multiplier_ext = NULL)
}
\arguments{
\item{penalty_type}{type of regularization for x. Default is 0 (Ridge). Can supply either a scalar value or vector with length equal to the number of variables in x.
\itemize{
   \item 0 = Ridge
   \item (0,1) = Elastic-Net
   \item 1 = Lasso / Quantile
}}

\item{quantile}{specifies quantile for predictors. Default of 0.5 reduces to lasso.}

\item{penalty_type_ext}{type of regularization for external data. See penalty_type for options. Default is 1 (lasso). Can supply either a scalar value or vector with length equal to the number of variables in external.}

\item{quantile_ext}{specifies quantile for external data. Default of 0.5 reduces to lasso.}

\item{num_penalty}{number of penalty values to fit in grid for x. Default is 20.}

\item{num_penalty_ext}{number of penalty values to fit in grid for external data. Default is 20.}

\item{penalty_ratio}{ratio between minimum and maximum penalty for x. Default is 1e-04 if \eqn{n > p} and 0.01 if \eqn{n <= p}.}

\item{penalty_ratio_ext}{ratio between minimum and maximum penalty for external data. Default is 1e-04 if \eqn{p > q} and 0.01 if \eqn{p <= q}.}

\item{user_penalty}{user-defined vector of penalty values to fit for x.}

\item{user_penalty_ext}{user-defined vector of penalty values to fit for external data.}

\item{custom_multiplier}{variable-specific penalty multipliers for x. Default is 1 for all variables. 0 is no penalization.}

\item{custom_multiplier_ext}{variable-specific penalty multipliers for external data. Default is 1 for all variables. 0 is no penalization.}
}
\description{
Defines regularization terms for predictor and external data in \code{\link{hierr}} fitting.
}
