% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierr.R
\name{hierr}
\alias{hierr}
\title{Fit hierarchical regularized regression model}
\usage{
hierr(x, y, external = NULL, unpen = NULL, family = c("gaussian"),
  penalty = definePenalty(0, 1), weights = NULL, standardize = c(TRUE,
  TRUE), intercept = c(TRUE, TRUE), control = list())
}
\arguments{
\item{x}{predictor design matrix of dimension \eqn{n x p}}

\item{y}{outcome vector of length \eqn{n}}

\item{external}{(optional) external data design matrix of dimension \eqn{p x q}}

\item{unpen}{(optional) unpenalized predictor design matrix}

\item{family}{error distribution for outcome variable}

\item{penalty}{specifies regularization object for x and external. See \code{\link{definePenalty}} for more details.}

\item{weights}{optional vector of observation-specific weights. Default is 1 for all observations.}

\item{standardize}{indicates whether x and/or external should be standardized. Default is c(TRUE, TRUE).}

\item{intercept}{indicates whether an intercept term is included for x and/or external. Default is c(TRUE, TRUE).}

\item{control}{specifies hierr control object. See \code{\link{hierr.control}} for more details.}
}
\description{
Fit hierarchical regularized regression model
}
