% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cvhierr.R
\name{plot.cv_hierr}
\alias{plot.cv_hierr}
\title{Plot k-fold cross-validation error grid}
\usage{
\method{plot}{cv_hierr}(x, ...)
}
\arguments{
\item{x}{A cv_hierr class object}

\item{...}{Additional graphics parameters}
}
\description{
Plot k-fold cross-validation error grid
}
