% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cvhierr.R
\name{plot.cvhierr}
\alias{plot.cvhierr}
\title{Plot k-fold cross-validation error grid}
\usage{
\method{plot}{cvhierr}(x, ...)
}
\arguments{
\item{x}{A cvhierr class object}

\item{...}{Additional graphics parameters}
}
\description{
Plot k-fold cross-validation error grid
}
