% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cvhierr.R
\name{predict.cvhierr}
\alias{predict.cvhierr}
\title{Predict function for "cvhierr" object}
\usage{
\method{predict}{cvhierr}(object, newdata = NULL, p = NULL,
  pext = NULL, type = c("response", "coefficients"), ...)
}
\arguments{
\item{object}{A \code{\link{cv_hierr}} object}

\item{newdata}{matrix with new X values}

\item{p}{vector of penalty values to apply to predictor variables}

\item{pext}{vector of penalty values to apply to external data variables}

\item{type}{type of prediction to make using the hierr model}

\item{...}{pass other arguments to hierr function (if needed)}
}
\description{
Predict coefficients or response in new data
}
