% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_hierr.R
\name{predict.hierr}
\alias{predict.hierr}
\title{Predict function for "hierr" object}
\usage{
\method{predict}{hierr}(object, newdata = NULL, penalty = NULL,
  penalty_ext = NULL, type = c("link", "response", "coefficients",
  "nonzero", "class"), ...)
}
\arguments{
\item{object}{A \code{\link{hierr}} object}

\item{newdata}{matrix with new X values}

\item{penalty}{vector of penalty values to apply to predictor variables}

\item{penalty_ext}{vector of penalty values to apply to external data variables}

\item{type}{type of prediction to make using the hierr model}
}
\description{
Predict coefficients or response in new data
}
