% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_tune_xrnet.R
\name{predict.tune_xrnet}
\alias{predict.tune_xrnet}
\title{Predict function for "tune_xrnet" object}
\usage{
\method{predict}{tune_xrnet}(object, newdata = NULL,
  newdata_fixed = NULL, p = "opt", pext = "opt",
  type = c("response", "coefficients", "link"), penalty = NULL, ...)
}
\arguments{
\item{object}{A \code{\link{tune_xrnet}} object}

\item{newdata}{matrix with new values for penalized variables}

\item{newdata_fixed}{matrix with new values for unpenalized variables}

\item{p}{vector of penalty values to apply to predictor variables.
Default is optimal value in tune_xrnet object.}

\item{pext}{vector of penalty values to apply to external data variables.
Default is optimal value in tune_xrnet object.}

\item{type}{type of prediction to make using the xrnet model, options include:
\itemize{
   \item coefficients
   \item response
   \item link (linear predictor)
}}

\item{penalty}{(optional) regularization object applied to original model object, only needed
if p or pext are not in the original path(s) computed. See \code{\link{define_penalty}} for
more information on regularization object.}

\item{...}{pass other arguments to xrnet function (if needed)}
}
\description{
Predict coefficients or response in new data
}
