% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_xrnet.R
\name{predict.xrnet}
\alias{predict.xrnet}
\title{Predict function for "xrnet" object}
\usage{
\method{predict}{xrnet}(object, newdata = NULL, newdata_fixed = NULL,
  p = NULL, pext = NULL, type = c("response", "coefficients",
  "link"), penalty = NULL, ...)
}
\arguments{
\item{object}{A \code{\link{xrnet}} object}

\item{newdata}{matrix with new values for penalized variables}

\item{newdata_fixed}{matrix with new values for unpenalized variables}

\item{p}{vector of penalty values to apply to predictor variables}

\item{pext}{vector of penalty values to apply to external data variables}

\item{type}{type of prediction to make using the xrnet model, options include
\itemize{
   \item coefficients
   \item response
   \item link (linear predictor)
}}

\item{penalty}{(optional) regularization object applied to original model object, only needed
if p or pext are not in the original path(s) computed. See \code{\link{define_penalty}} for
more information on regularization object.}

\item{...}{pass other arguments to xrnet function (if needed)}
}
\description{
Predict coefficients or response in new data
}
