% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x_linear}
\alias{x_linear}
\title{Simulated example data for hierarchical regularized linear regression}
\format{
A matrix with 100 rows and 50 variables
}
\usage{
x_linear
}
\description{
Simulated example data for hierarchical regularized linear regression
}
\keyword{datasets}
