% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_xrnet.R
\name{coef.xrnet}
\alias{coef.xrnet}
\title{Get coefficient estimates from "xrnet" model object}
\usage{
\method{coef}{xrnet}(object, p = NULL, pext = NULL, penalty = NULL,
  ...)
}
\arguments{
\item{object}{A \code{\link{xrnet}} object}

\item{p}{vector of penalty values to apply to predictor variables.}

\item{pext}{vector of penalty values to apply to external data variables.}

\item{penalty}{(optional) regularization object applied to original model object, only needed
if p or pext are not in the original path(s) computed. See \code{\link{define_penalty}} for
more information on regularization object.}

\item{...}{pass other arguments to xrnet function (if needed)}
}
\value{
A list with coefficient estimates at each of the requested penalty combinations
}
\description{
Returns coefficents from 'xrnet' model or refit model to estimate
coefficients not computed in original path(s) of model object.
}
\examples{
data(GaussianExample)

fit_xrnet <- xrnet(
    x = x_linear,
    y = y_linear,
    external = ext_linear,
    family = "gaussian"
)

lambda1 <- fit_xrnet$penalty[10]
lambda2 <- fit_xrnet$penalty[10]
\dontrun{
coef_xrnet <- coef(
    fit_xrnet,
    p = lambda1,
    pext = lambda2,
)
}

}
