% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvhierr.R
\name{cv_hierr}
\alias{cv_hierr}
\title{k-fold cross-validation for hierarchical regularized regression}
\usage{
cv_hierr(x, y, external = NULL, unpen = NULL, family = c("gaussian",
  "binomial"), penalty = define_penalty(), weights = NULL,
  standardize = c(TRUE, TRUE), intercept = c(TRUE, FALSE),
  loss = c("mse", "mae", "auc"), nfolds = 5, foldid = NULL,
  parallel = FALSE, control = list())
}
\arguments{
\item{x}{predictor design matrix of dimension \eqn{n x p}, matrix options include:
\itemize{
   \item matrix
   \item big.matrix
   \item filebacked.big.matrix
   \item sparse matrix (dgCMatrix)
}}

\item{y}{outcome vector of length \eqn{n}}

\item{external}{(optional) external data design matrix of dimension \eqn{p x q}, matrix options include:
\itemize{
    \item matrix
    \item sparse matrix (dgCMatrix)
}}

\item{unpen}{(optional) unpenalized predictor design matrix}

\item{family}{error distribution for outcome variable, options include:
\itemize{
    \item gaussian
    \item binomial
}}

\item{penalty}{specifies regularization object for x and external. See \code{\link{define_penalty}} for more details.}

\item{weights}{optional vector of observation-specific weights. Default is 1 for all observations.}

\item{standardize}{indicates whether x and/or external should be standardized. Default is c(TRUE, TRUE).}

\item{intercept}{indicates whether an intercept term is included for x and/or external. Default is c(TRUE, FALSE).}

\item{loss}{loss function for cross-validation. Options include:
\itemize{
   \item mse (Mean Squared Error)
   \item mae (Mean Absolute Error)
   \item auc (Area under the curve)
}}

\item{nfolds}{number of folds for cross-validation. Default is 5.}

\item{foldid}{(optional) vector that identifies user-specified fold for each observation. If NULL, folds are automatically generated.}

\item{parallel}{use \code{foreach} function to fit folds in parallel if TRUE, must register cluster (\code{doParallel}) before using.}

\item{control}{specifies hierr control object. See \code{\link{hierr.control}} for more details.}
}
\value{
A list of class \code{cvhierr} with components
\item{cv_mean}{mean cross-validated error for each penalty combination. Object returned is
a vector if there is no external data (external = NULL) and matrix if there is external data.}
\item{cv_sd}{estimated standard deviation for cross-validated errors}
\item{opt_loss}{the value for the optimal cross-validated error}
\item{opt_penalty}{first-level penalty value that achieves the optimal loss}
\item{opt_penalty_ext}{second-level penalty value that achieves the optimal loss (if external data is present)}
\item{fitted_model}{fitted hierr object using all data, see \code{\link{hierr}} for details of object}
}
\description{
k-fold cross-validation for hierarchical regularized regression \code{\link{hierr}}
}
