% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierr.R
\name{definePenalty}
\alias{definePenalty}
\title{Define regularization object for predictor and external data}
\usage{
definePenalty(penalty_type, penalty_type_ext, num_penalty = 50,
  num_penalty_ext = 50, penalty_ratio = NULL, penalty_ratio_ext = NULL,
  user_penalty = NULL, user_penalty_ext = NULL, custom_multiplier = NULL,
  custom_multiplier_ext = NULL)
}
\arguments{
\item{penalty_type}{type of regularization for x. Default is 0 (Ridge).
\itemize{
   \item 0 = Ridge
   \item (0,1) = Elastic-Net
   \item 1 = Lasso
   \item 2 = Quantile
}}

\item{penalty_type_ext}{type of regularization for external data. See penalty_type for options. Default is 1 (Lasso).}

\item{num_penalty}{number of penalty values to fit in grid for x. Default is 50.}

\item{num_penalty_ext}{number of penalty values to fit in grid for external data. Default is 50.}

\item{penalty_ratio}{ratio between minimum and maximum penalty for x. Default is 1e-04 if \eqn{n > p} and 0.01 if \eqn{n <= p}.}

\item{penalty_ratio_ext}{ratio between minimum and maximum penalty for external data. Default is 1e-04 if \eqn{p > q} and 0.01 if \eqn{p <= q}.}

\item{user_penalty}{user-defined vector of penalty values to fit for x.}

\item{user_penalty_ext}{user-defined vector of penalty values to fit for external data.}

\item{custom_multiplier}{variable-specific penalty multipliers for x. Default is 1 for all variables.}

\item{custom_multiplier_ext}{variable-specific penalty multipliers for external data. Default is 1 for all variables.}
}
\description{
Defines regularization terms for predictor and external data in \code{\link{hierr}} fitting.
}
