% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierr.R
\name{hierr}
\alias{hierr}
\title{Fit hierarchical regularized regression model}
\usage{
hierr(x, y, external = NULL, unpen = NULL, family = c("gaussian"),
  penalty = definePenalty(), weights = NULL, standardize = c(TRUE,
  TRUE), intercept = c(TRUE, TRUE), control = list())
}
\arguments{
\item{x}{predictor design matrix of dimension \eqn{n x p}}

\item{y}{outcome vector of length \eqn{n}}

\item{external}{(optional) external data design matrix of dimension \eqn{p x q}}

\item{unpen}{(optional) unpenalized predictor design matrix}

\item{family}{error distribution for outcome variable}

\item{penalty}{specifies regularization object for x and external. See \code{\link{definePenalty}} for more details.}

\item{weights}{optional vector of observation-specific weights. Default is 1 for all observations.}

\item{standardize}{indicates whether x and/or external should be standardized. Default is c(TRUE, TRUE).}

\item{intercept}{indicates whether an intercept term is included for x and/or external. Default is c(TRUE, TRUE).}

\item{control}{specifies hierr control object. See \code{\link{hierr.control}} for more details.}
}
\value{
A list of class \code{hierr} with components
\item{beta0}{matrix of first-level intercepts indexed by penalty values}
\item{betas}{3-dimensional array of first-level penalized coefficients indexed by penalty values}
\item{gammas}{3-dimensional array of first-level non-penalized coefficients indexed by penalty values}
\item{alpha0}{matrix of second-level intercepts indexed by penalty values}
\item{alphas}{3-dimensional array of second-level external data coefficients indexed by penalty values}
\item{penalty}{vector of first-level penalty values}
\item{penalty_ext}{vector of second-level penalty values}
\item{penalty_type}{type of penalty applied to first-level predictors}
\item{quantile}{quantile for penalty on first-level predictors}
\item{penalty_type_ext}{type of penalty applied to second-level external data}
\item{quantile_ext}{quantile for penalty on second-level external data}
\item{penalty_ratio}{ratio between minimum and maximum penalty (predictors)}
\item{penalty_ratio_ext}{ratio between minimum and maximum penalty (external data)}
\item{deviance}{fraction of deviance explaned, \eqn{2 x (loglike(saturated model) - loglike(model))}}
\item{nlp}{total number of passes over data}
\item{custom_mult}{vector of variable-specific penalty multipliers for predictors}
\item{custom_mult_ext}{vector of variable-specific penalty multipliers for external data}
}
\description{
Fits hierarchical regularized regression model that enables the incorporation of external data
for the predictor variables. Both the predictor variables and external data can be regularized
by the most common penalties (lasso, ridge, elastic net) and we have included an addtional "quantile" penalty.
Solutions are computed across a two-dimensional grid of penalties (a separate penalty path is computed
for the predictors and external variables). Currently support linear regression, future extensions to
a GLM framework will be implemented in the next major update.
}
