% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tune_xrnet.R
\name{plot.tune_xrnet}
\alias{plot.tune_xrnet}
\title{Plot k-fold cross-validation error grid}
\usage{
\method{plot}{tune_xrnet}(x, ...)
}
\arguments{
\item{x}{A tune_xrnet class object}

\item{...}{Additional graphics parameters}
}
\description{
Plot k-fold cross-validation error grid
}
